[index] implement diffing for V2 via IndexV2DiffStreamProcessor
This uses hacky reflection and doesn't properly diff everything it could, but just what we need for the DB

This commit also sneaks in some last minute change requests for the indexV2 structure

+++ index/src/androidMain/kotlin/org/fdroid/index/v1/IndexV1StreamProcessor.kt
@@ -91,7 +91,7 @@ public class IndexV1StreamProcessor(
                 categories = emptyMap(),
                 releaseChannels = emptyMap()
             )
-            indexStreamReceiver.receive(repoV2, repo.version, certificate)
+            indexStreamReceiver.receive(repoV2, repo.version.toLong(), certificate)
         }
 
         private fun deserializeRequests(decoder: JsonDecoder, index: Int) {
